refactor(server): rename reaction/ → trigger/ module (ADR-0046 slice 9c/n)
Fulfills ADR-0046's naming-consistency commitment. The module
structure now matches the public-facing term developers use in
[[action.triggers]] spec blocks.

Scope:
- `git mv` of crates/temper-server/src/reaction/ → src/trigger/
  (8 files; git rename tracking preserved across the history).
- Bulk import-path update across the workspace:
  * `crate::reaction::` → `crate::trigger::`
  * `temper_server::reaction::` → `temper_server::trigger::`
  * `pub mod reaction;` → `pub mod trigger;`
  14 files touched.

Not changed in this commit (deliberately scoped):
- Type names (ReactionRule, ReactionDispatcher, ReactionRegistry,
  ReactionResult, ReactionGuard, ReactionTarget, ReactionTrigger)
  remain. Renaming these would cascade across ~50 more call sites
  without changing behavior. A follow-up slice can rename the types
  if the inconsistency proves to matter in practice; for now the
  module path is the visible ADR commitment.
- Integration test filenames (reaction_e2e_prod.rs, reaction_cascade.rs)
  left as-is so their git history stays intact; they exercise the
  renamed module correctly.

Workspace builds green. All per-crate lib tests pass (242 + 94 + 51 +
301 + 168 = 856 tests). Integration tests pass: trigger_e2e_prod (1),
reaction_e2e_prod (4), reaction_cascade (12).
