+++ crates/temper-server/tests/reaction_e2e_prod.rs
@@ -12,7 +12,7 @@ use std::sync::Arc;
 use temper_runtime::ActorSystem;
 use temper_runtime::tenant::TenantId;
 use temper_server::ServerState;
-use temper_server::reaction::registry::parse_reactions;
+use temper_server::trigger::registry::parse_reactions;
 use temper_server::registry::SpecRegistry;
 use temper_server::request_context::AgentContext;
 use temper_spec::csdl::parse_csdl;
